requirements.txt
streamlit>=1.31.0
pandas>=2.1.0
numpy>=1.24.0
yfinance>=0.2.36
altair>=5.2.0
google-genai>=0.3.0
python-dotenv>=1.0.1